Backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
langchain==0.0.300
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf==3.16.2
python-multipart==0.0.6
python-dotenv==1.0.0
huggingface-hub==0.17.3